<feature>Add jwtMiddleware
